rename NS
