+++ src/User/Infra/Console/Command/CreateUserCommand.php
@@ -5,7 +5,7 @@ declare(strict_types=1);
 namespace MsgPhp\User\Infra\Console\Command;
 
 use MsgPhp\Domain\Factory\DomainObjectFactoryInterface;
-use MsgPhp\Domain\Infra\Console\ContextBuilder\ContextBuilderInterface;
+use MsgPhp\Domain\Infra\Console\Context\ContextFactoryInterface;
 use MsgPhp\Domain\Message\{DomainMessageBusInterface, MessageDispatchingTrait, MessageReceivingInterface};
 use MsgPhp\User\Command\CreateUserCommand as CreateUserDomainCommand;
 use MsgPhp\User\Event\UserCreatedEvent;
@@ -29,7 +29,7 @@ final class CreateUserCommand extends Command implements MessageReceivingInterfa
     /** @var StyleInterface */
     private $io;
 
-    public function __construct(DomainObjectFactoryInterface $factory, DomainMessageBusInterface $bus, ContextBuilderInterface $contextBuilder)
+    public function __construct(DomainObjectFactoryInterface $factory, DomainMessageBusInterface $bus, ContextFactoryInterface $contextBuilder)
     {
         $this->factory = $factory;
         $this->bus = $bus;
