+++ src/UserBundle/DependencyInjection/Extension.php
@@ -124,18 +124,18 @@ final class Extension extends BaseExtension implements PrependExtensionInterface
             $loader->load('console.php');
 
             $container->getDefinition(ConsoleInfra\Command\AddUserRoleCommand::class)
-                ->setArgument('$contextBuilder', ContainerHelper::registerAnonymous($container, BaseConsoleInfra\ContextBuilder\ClassContextBuilder::class, true)
+                ->setArgument('$contextBuilder', ContainerHelper::registerAnonymous($container, BaseConsoleInfra\Context\ClassContextFactory::class, true)
                     ->setArgument('$class', Entity\UserRole::class)
-                    ->setArgument('$flags', BaseConsoleInfra\ContextBuilder\ClassContextBuilder::REUSE_DEFINITION));
+                    ->setArgument('$flags', BaseConsoleInfra\Context\ClassContextFactory::REUSE_DEFINITION));
 
             $container->getDefinition(ConsoleInfra\Command\CreateUserCommand::class)
-                ->setArgument('$contextBuilder', ContainerHelper::registerAnonymous($container, BaseConsoleInfra\ContextBuilder\ClassContextBuilder::class, true)
+                ->setArgument('$contextBuilder', ContainerHelper::registerAnonymous($container, BaseConsoleInfra\Context\ClassContextFactory::class, true)
                     ->setArgument('$class', Entity\User::class));
 
             $container->getDefinition(ConsoleInfra\Command\ChangeUserCredentialCommand::class)
-                ->setArgument('$contextBuilder', ContainerHelper::registerAnonymous($container, BaseConsoleInfra\ContextBuilder\ClassContextBuilder::class, true)
+                ->setArgument('$contextBuilder', ContainerHelper::registerAnonymous($container, BaseConsoleInfra\Context\ClassContextFactory::class, true)
                     ->setArgument('$class', CredentialInterface::class)
-                    ->setArgument('$flags', BaseConsoleInfra\ContextBuilder\ClassContextBuilder::ALWAYS_OPTIONAL | BaseConsoleInfra\ContextBuilder\ClassContextBuilder::NO_DEFAULTS));
+                    ->setArgument('$flags', BaseConsoleInfra\Context\ClassContextFactory::ALWAYS_OPTIONAL | BaseConsoleInfra\Context\ClassContextFactory::NO_DEFAULTS));
 
             ContainerHelper::removeIf($container, !$container->has(Command\Handler\AddUserRoleHandler::class), [
                 ConsoleInfra\Command\AddUserRoleCommand::class,
